+++ migrate.php
@@ -708,6 +708,7 @@ $newSettings = [
     'og_image_default'        => '',
     'site_favicon'            => '',
     'site_logo'               => '',
+    'site_profile'            => guessSiteProfileKey(),
     'active_theme'            => defaultThemeName(),
     // Cookie lišta
     'cookie_consent_enabled'  => '0',